requirements.txt
aioboto3==15.5.0
aiobotocore==2.25.1
aioconsole==0.8.2
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioitertools==0.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.14.3
boto3==1.40.61
botocore==1.40.61
certifi==2025.11.12
cffi==2.0.0
cryptography==46.0.3
deprecation==2.1.0
discord.py==2.6.4
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jmespath==1.0.1
multidict==6.7.0
packaging==25.0
pillow==12.0.0
postgrest==2.25.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
realtime==2.25.0
s3transfer==0.14.0
scapi==3.1.4
six==1.17.0
soupsieve==2.8
storage3==2.25.0
StrEnum==0.4.15
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
